+++ activity/brightpay/nextjs/changelog.md
@@ -3982,3 +3982,14 @@
 
 ---
 
+# Brightpay Development - 2025-01-01
+
+**Commit #1**
+
+- Repository: brightpay/brightclinic
+- Type: healthcare-platform
+- Language: nextjs
+- Month: 1/2025
+
+---
+